requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.0
python-docx>=0.8.11
typing-extensions>=4.0.0
streamlit-extras>=0.2.0
streamlit-pdf-viewer>=0.0.23
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
scikit-learn>=1.1.0
nltk>=3.7
requests>=2.28.0
Pillow>=9.0.0
dask>=2023.3.0
distributed>=2023.3.0
lightgbm>=4.1.0
tensorflow>=2.15.0
imbalanced-learn>=0.11.0
imblearn>=0.0 